refactor: share structural factory detection
