+++ packages/ts-transformers/src/transformers/expression-site-policy.ts
@@ -25,9 +25,13 @@ import {
   isJsxLocalRewriteContainer,
 } from "./expression-rewrite/emitters/compute-wrap-invariants.ts";
 import type { AnalyzeFn } from "./expression-rewrite/types.ts";
-import { isOpaqueRefType } from "./opaque-ref/opaque-ref.ts";
 import { classifyOpaquePathTerminalCall } from "./opaque-roots.ts";
 import type { ExpressionContainerKind } from "./expression-site-types.ts";
+import {
+  isPatternFactoryCalleeExpression,
+  isStructuralReactiveFactoryExpression,
+  returnsOpaqueRefResult,
+} from "./structural-reactive-factory.ts";
 
 interface ExpressionSiteCallRootPolicyInfo {
   readonly reactiveContext: ReactiveContextInfo;
@@ -209,179 +213,6 @@ function getLeftmostMemberBase(expression: ts.Expression): ts.Expression {
     current = current.expression;
   }
   return current;
-}
-
-function isPatternFactoryCalleeExpression(
-  expression: ts.Expression,
-  checker: ts.TypeChecker,
-): boolean {
-  const target = unwrapExpression(expression);
-
-  try {
-    const type = checker.getTypeAtLocation(target);
-    const signatures = checker.getSignaturesOfType(type, ts.SignatureKind.Call);
-    if (signatures.length === 0) {
-      return false;
-    }
-
-    const propertyNames = new Set(
-      type.getProperties().map((property) => property.getName()),
-    );
-    if (
-      !propertyNames.has("argumentSchema") ||
-      !propertyNames.has("resultSchema") ||
-      propertyNames.has("with")
-    ) {
-      return false;
-    }
-
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function returnsOpaqueRefResult(
-  expression: ts.CallExpression,
-  checker: ts.TypeChecker,
-): boolean {
-  try {
-    const type = checker.getTypeAtLocation(expression);
-    return isOpaqueRefType(type, checker);
-  } catch {
-    return false;
-  }
-}
-
-function getAliasedSymbol(
-  symbol: ts.Symbol,
-  checker: ts.TypeChecker,
-): ts.Symbol {
-  if (!(symbol.flags & ts.SymbolFlags.Alias)) {
-    return symbol;
-  }
-
-  try {
-    return checker.getAliasedSymbol(symbol);
-  } catch {
-    return symbol;
-  }
-}
-
-function getReturnedExpression(
-  declaration: ts.Declaration,
-): ts.Expression | undefined {
-  if (
-    ts.isFunctionDeclaration(declaration) ||
-    ts.isMethodDeclaration(declaration) ||
-    ts.isFunctionExpression(declaration) ||
-    ts.isArrowFunction(declaration)
-  ) {
-    if (!declaration.body) {
-      return undefined;
-    }
-    if (ts.isBlock(declaration.body)) {
-      if (declaration.body.statements.length !== 1) {
-        return undefined;
-      }
-      const [statement] = declaration.body.statements;
-      return ts.isReturnStatement(statement) ? statement.expression : undefined;
-    }
-    return declaration.body;
-  }
-
-  if (!ts.isVariableDeclaration(declaration) || !declaration.initializer) {
-    return undefined;
-  }
-
-  const initializer = unwrapExpression(declaration.initializer);
-  if (
-    ts.isArrowFunction(initializer) ||
-    ts.isFunctionExpression(initializer)
-  ) {
-    return getReturnedExpression(initializer);
-  }
-
-  if (
-    ts.isIdentifier(initializer) ||
-    ts.isCallExpression(initializer) ||
-    ts.isPropertyAccessExpression(initializer)
-  ) {
-    return initializer;
-  }
-
-  return undefined;
-}
-
-function isStructuralReactiveFactoryExpression(
-  expression: ts.Expression,
-  checker: ts.TypeChecker,
-  seenSymbols = new Set<ts.Symbol>(),
-): boolean {
-  const target = unwrapExpression(expression);
-
-  if (ts.isCallExpression(target)) {
-    if (returnsOpaqueRefResult(target, checker)) {
-      return true;
-    }
-
-    if (isPatternFactoryCalleeExpression(target.expression, checker)) {
-      return true;
-    }
-
-    const callKind = detectCallKind(target, checker);
-    if (callKind) {
-      switch (callKind.kind) {
-        case "builder":
-        case "derive":
-        case "cell-factory":
-        case "cell-for":
-        case "wish":
-        case "generate-text":
-        case "generate-object":
-        case "pattern-tool":
-        case "runtime-call":
-          return true;
-        default:
-          break;
-      }
-    }
-
-    return isStructuralReactiveFactoryExpression(
-      target.expression,
-      checker,
-      seenSymbols,
-    );
-  }
-
-  if (
-    !ts.isIdentifier(target) &&
-    !ts.isPropertyAccessExpression(target)
-  ) {
-    return false;
-  }
-
-  const symbol = checker.getSymbolAtLocation(target);
-  if (!symbol) {
-    return false;
-  }
-
-  const resolvedSymbol = getAliasedSymbol(symbol, checker);
-  if (seenSymbols.has(resolvedSymbol)) {
-    return false;
-  }
-  seenSymbols.add(resolvedSymbol);
-
-  const declarations = resolvedSymbol.getDeclarations() ?? [];
-  return declarations.some((declaration) => {
-    const returnedExpression = getReturnedExpression(declaration);
-    return !!returnedExpression &&
-      isStructuralReactiveFactoryExpression(
-        returnedExpression,
-        checker,
-        seenSymbols,
-      );
-  });
 }
 
 function classifyCallExpressionRoot(
